requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
Authlib==1.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
ctransformers==0.2.27
dataclasses-json==0.6.3
diskcache==5.6.3
et-xmlfile==1.1.0
fastapi==0.109.2
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.1
grpcio==1.60.1
grpcio-health-checking==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
langchain==0.0.338
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.78
llama_cpp_python==0.2.43
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
protobuf==4.25.3
psutil==5.9.7
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
textile==4.0.2
torch==2.2.0
tqdm==4.66.1
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.27.1
validators==0.22.0
weaviate-client==3.25.2
webencodings==0.5.1
Werkzeug==2.3.7
XlsxWriter==3.1.9
yarl==1.9.4
